% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_results-class.R
\name{is_sens_results}
\alias{is_sens_results}
\title{Check if object is a sens_results}
\usage{
is_sens_results(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Logical.
}
\description{
Check if object is a sens_results
}
