% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_results-class.R
\name{print.sens_results}
\alias{print.sens_results}
\title{Print method for sens_results}
\usage{
\method{print}{sens_results}(x, ...)
}
\arguments{
\item{x}{A \code{sens_results} object.}

\item{...}{Further arguments (unused).}
}
\description{
Print method for sens_results
}
