% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_results-class.R
\name{summary.sens_results}
\alias{summary.sens_results}
\title{Summary method for sens_results}
\usage{
\method{summary}{sens_results}(object, ...)
}
\arguments{
\item{object}{A \code{sens_results} object.}

\item{...}{Further arguments (unused).}
}
\description{
Summary method for sens_results
}
