% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_results-class.R
\name{plot.sens_results}
\alias{plot.sens_results}
\title{Plot method for sens_results}
\usage{
\method{plot}{sens_results}(x, ...)
}
\arguments{
\item{x}{A \code{sens_results} object.}

\item{...}{Passed to \code{visualize_sensitivity()}.}
}
\value{
Invisibly returns \code{x}, the unchanged \code{sens_results}
  object. The method is called for its side effect of drawing a sensitivity plot.
}
\description{
Plot method for sens_results
}
